+++ News.md
@@ -1,5 +1,7 @@
 # News
 
+- `TreeSitter::Node` is enumerable.
+
 # v1.3.0
 
 - Query Predicates landed. See https://tree-sitter.github.io/tree-sitter/using-parsers#predicates.